Rename django test app to be more descriptive
