+++ tests/django_logging/test_app/views.py
